khasanovavera/requirements.txt
dotenv == 0.9.9
telebot == 0.0.5
langchain>=1.2.10,<1.3.0                
langchain-community>=0.4.0,<0.5.0       
langchain-huggingface>=1.2.0,<1.3.0     
langchain-classic>=1.0.0,<1.1.0         
langchain-openai>=1.0.0,<1.2.0
langgraph>=1.0.8,<1.1.0
faiss-cpu>=1.8.0                        
langchain-chroma>=0.2.0,<0.3.0
sentence-transformers>=5.2.0,<5.3.0     
transformers>=4.48.0,<5.0.0             
torch>=2.4.0                            
duckduckgo-search>=6.0.0
ddgs
tavily-python>=0.4.0                    
pypdf>=5.0.0
pdfplumber>=0.11.0
jupyter>=1.0.0
ipywidgets>=8.1.0
matplotlib>=3.9.0
numpy>=2.0.0                            
pandas>=2.2.0
tqdm>=4.66.0
python-dotenv>=1.0.0
seaborn>=0.13.0
scikit-learn>=1.5.0